Don't try to update the dots buffer if there are no dots
Previously if there are no dots then Rig would still try to map a
zero-sized range of the dots buffer that it allocates. When that fails
it would then try to allocate a zero-sized buffer and update a sub
region of the attribute buffer from that. Both of these will probably
generate GL errors. This patch makes it just leave the buffer alone
when there are no dots.