Add a RutBin
This adds a RutBin widget which is a container to contain a single
child. The child can be position with padding around all four of its
sides. The position of the child when there is more space than it
needs can be customised for each axis to either be at the left/top,
center, bottom/right or to expand to fill the space.