Use a RutBin for the search widget instead of a fixed size
The search wiget for the assets is now placed inside a RutBin which is
directly placed in the stack for the left side of the split view. That
way the bin will always have the full size of the split. Some padding
is set on the bin to position the search widget and it is set to
expand along the x axis. That way the search widget will automatically
change size when the split is moved instead of having an arbitrary
fixed size.