Add support for setting the mouse cursor image in SDL2
There is now a new function called rut_shell_set_cursor which
temporarily changes the cursor image. This takes an enum of predefined
images. When using the SDL2 backend, this will set the cursor to the
corresponding system cursor.

This function is called by the shell in response to every mouse motion
event when no other handler calls it. The idea is that a widget can
change the cursor whenever the pointer enters its input region and the
shell will then automatically change it back when the pointer leaves
that input region.