split-view: Use a constant to specify the size of the grab area
Previously the size of the grab error was hard coded in the code.
There was a discrepency between the size of the painted area and the
size of the actual input region for the grab. The input region was one
pixel bigger than the area allocated for it so it would end up
overlapping with the area for one of the children. This has been fixed
so that both sizes are now the same.