split-view: Set the child1 transform before the size
The code for handling a size change on the main editor area assumes
that the transform for the area is also correct in order to position
the window-relative viewport. This was not previously correct because
the transform was set after setting the size so it would end up using
the old position for the viewport. This patch just swaps the order so
that it works.