Add a GTK/X11 backend for RutShell
RutShell now has a backend to get events via GDK. It will force Cogl
to use an X11 winsys. It will then ask Cogl for the X window ID and
create a foreign GDK window in order to get events.

The backend is optional with a configure option called --enable-gtk.