Add rut_shell_add_onscreen so that the shell can track onscreens
It's useful for the RutShell to be aware of all onscreen windows that
are used in the application so that it can eventually report events to
the right onscreen when multiple windows are used. This adds a
rut_shell_add_onscreen() function which adds the onscreen to a list.
The onscreen is automatically removed from the list when it is
destroyed.