Add a function to get the CoglOnscreen which corresponds to an event
Eventually Rig will want to support multiple onscreen windows and
we'll need to determine which onscreen relates to which event. This
adds a function called rut_input_event_get_onscreen which returns the
onscreen. On SDL2 it will walk the list of known onscreen buffers to
find one which corresponds to an SDL window with the same ID. On all
other backends there can only be one window so it will just return
that.