Add a GApplication subclass with a menu when running with Gtk
When using the GTK backend Rig will now create a GApplication and
advertise a menu over DBus so that it can appear in the gnome shell
panel.

The RutShell is now able to manage starting the GApplication if one is
installed when the shell is run instead of starting the gtk main loop.
There is also a new function to query the GdkWindow from a
CoglOnscreen via the shell so that Rig can make the dialog box for the
file chooser be modal for the Rig window.