Query rectangle tex parameters when creating a foreign texture on GL3
Previously when creating a foreign rectangle texture it would ignore
the passed in texture information and query the texture directly when
using COGL_DRIVER_GL. However this should also work for
COGL_DRIVER_GL3. This patch changes it to check the private feature
flags for the texture querying feature instead of directly checking
the driver value.

Reviewed-by: Robert Bragg <robert@linux.intel.com>