Use and verify a signature on the automatic updates on OS X
In the tools directory a tool called rig-check-signature now gets
built. This uses libcrypto to verify the downloaded update against a
known public key. The public key is stored in the git repo. The
tarballs for the update are now prefixed with a 128 byte signature
which signs the SHA256 digest of the remainder of the file. If the
signature matches the tool exits with 0 status and writes the
remainder of the file out to a given location.

The automatic update script on OS X is now configured to use the tool
before moving the extracted data to the update filename. That way it
should only use trusted files.

The build scripts for OS X now also build a static library of
libcrypto so that it can be linked to by the tool.