box-layout: improve packing algorithm
This ports some of the allocation code from gtkbox.c to
rut-box-layout.c, primarily so that we taking into account the size of
the box-layout itself.

This also adds support for right-to-left and bottom-to-top packing;
setting a fixed spacing between children; homogeneous or non-homogeneous
packing of children; consideration for the users text-direction; setting
children to expand or not.