sdl: Wakeup the event blocking immediately if necessary
It is expected that cogl_sdl_idle() will be called from the
application immediately before blocking in SDL_WaitEvent. However,
dispatching the onscreen events may cause more events to be queued. If
that happens we need to make sure the blocking returns immediately.
This patch makes it post the dummy event that the application chose in
order to make that happen.

Reviewed-by: Robert Bragg <robert@linux.intel.com>