Add an initial rig-view widget
This adds an initial rig-view widget that is intended to provide a
common shell for the modular views that make up the Rig UI, providing a
place for toolbar items and then a bin to hold the main contents of the
view. The plan is to also add a drop-down widget to the toolbar for
switching the view type.