factor out RutCameraView
This factors out all of the high-level code for painting, picking and
interacting with a scene into a RutCameraView widget.

The intention is that we should be able to show multiple views of the
scene at the same time, from different angles and so this is the first
step to enabling that.