camera: Add suspend/resume api
This adds rut_camera_suspend/resume functions that can be used to
temporarily stop an in-progress frame allowing any number of other
cameras to be used before resuming a camera frame.

While a camera is suspended it is currently expected that the camera
transforms will not be modified and we currently assert that this is the
case on resume. Potentially we could change the implementation to
explicitly copy the viewport and projection so we could allow camera
modifications (perhaps in preparation for a later frame).

Note: We had originally considered a _paint_context_push/pop_camera type
of api to support temporarily switching to a different camera but here
are some of the reasons we decided against that:
- we wanted to contain any code responsible for saving/restoring
  modelview, projection and viewport state in rut-camera.c
- we have use cases where we want to switch to rendering with multiple
  different cameras before returning to the original and those cameras
  aren't related hierarchically but instead sequentially and so having
  to _push() and _pop() each of them sequentially would end up doing
  redundant work to restore state for each _pop().
- it's not obvious if a _push() api would implicitly _flush() the
  specified camera's state or just save it to be restored in the _pop().