ui-viewport: remove x/y expand properties
To minimize the number of different ways we can control how widgets
should be expanded this removes the x/y expand properties from
RutUIViewport and it now expands according to whether panning is enabled
for each axis.

If x-axis panning is not enabled then we expand the sync widget on
the x-axis, and similarly if y-axis panning is not enabled we will
expand the sync widget on the y-axis.

If you don't want a widget expanded on the axis for which panning is
disabled then the widget can be put into a RutBin which lets you
also control alignment as well as expand behaviour.