rut-ui-viewport: be smarter about allocating sync widget
NB: when a sync widget has been specified for a ui-viewport then
depending on whether x/y_panning is enabled then the sync widget may be
expanded to fill the axis for which panning is disabled.

This improves the logic for allocating the sync widget and determining
the document size from the sync widget by querying the sync widget's
preferred size according to what constraints are implied by the
x/y_panning properties.  If for example we know that x_panning is
disabled and therefore the sync widget will be expanded to the width of
the viewport then we'll query the size of the sync-widget by passing a
for_width to _get_preferred_height. If y_panning is disabled then we
will pass a for_height to _get_preferred_width instead.

We will also re-query the preferred size of the sync-widget if it turns
out that we need to use a scrollbar that changes how much the
sync-widget will be expanded by.