stack: don't queue alloc when removing last child
there's no need to queue an allocation if a stack has no children.
Note: at some point it would be good to re-work the RutStack so that we
have a rut_stack_remove api and don't rely on the graphable callbacks
for removing child state. Relying on the graphable child_removed
callback makes it more awkward to avoid redundantly queuing
re-allocations when removing all the children when freeing the stack.