scroll-bar: strip down to just render the handle
This simplifies scrollbars so we only render a minimalist handle, and
when instantiating a scroll bar you now need to specify a handle colour.
Optionally handles can also have a second colour set on them so that the
colour will change between the start and end positions. Graduating the
colour can be useful when the scrollbar has a gradient background to
make sure it always contrasts with the background.