Adds initial RutFold widget
This adds a simple container for a single child which shows a label
above the child and a fold icon. When the top area is clicked then the
widget folds which updates the fold icon and either hides/shows the
given child widget depending on whether it is folding up or down.