Remove the GTK backend and only use GTK for the menu support
The GTK backend no longer works with the latest git master of GTK.
Something seems to be stopping it from receiving mouse motion events.
Perhaps the motion events in GTK are now collapsed between frames and
we are not getting them because GTK isn't aware that we are drawing
anything.

The GTK backend was also suffering from other problems which seem to
be due to GDK not really supporting having foreign toplevel windows
very well.

Unless we can get a proper GDK winsys in Cogl to integrate it
properly in the meantime it seems simpler to just continue using the
SDL backend on Linux. The menu and the dialog box can still work
because the RigApplication can query the X window from SDL. The file
chooser dialog box is made transient by explicitly using an Xlib call
with the window IDs of the two corresponding windows.

There doesn't seem to be a way in either SDL or GTK to get it to use a
foreign X Display. This means that Rig will end up with two
connections to X. Although that is a bit ridiculous I don't think it
will cause any problems in practice.