mingw: Update to the latest version of glib.
Cogl now requires at least version 2.32.0 of GLib. There are no
pre-built binaries for this from tml so we have to build glib from
source. This also requires libffi. The binary packages for libffi and
zlib don't install a pc file so the script just manually generates a
fake one.

Reviewed-by: Robert Bragg <robert@linux.intel.com>