rpc: Add way to get RigData in service implementation
This adds api that lets us associate a private pointer with a
PB_RPC_ServerConnection, and then later when we receive requests then
the functions implementing a service api can extract that private
pointer out from the closure_data argument.

Note: this is not an ideal approach because the
rig_pb_rpc_closure_get_connection_data() can't be type safe without us
changing the prototype of the service implementation functions which
would mean changing the codegen done by protoc-c. For now we are
avoiding importing protoc-c into rig, but later this could be
worthwhile.