load-save: Change to use binary protocol buffer format
This adds support for loading and saving UIs using the binary protocol
buffer format instead of xml.

Since we are aiming to use the protocol buffer format for our slave
device networking protocol it makes sense to only maintain one
serialization scheme.

The protocol buffer format should also be considerably faster to parse
than an XML file so this should also help to optimize application load
times.

For reference it's possible to dump the binary format into a human
readable form by running:

  $ protoc-c --decode=rig.UI rig.proto < binary-ui.rig

If you were to modify the human readable format then you convert to a
binary form by running:

  $ protoc-c --encode=rig.UI rig.proto < text-ui.rig > binary-ui.rig

For initial compatibility this patch preserves the xml file loading
and saving, though the plan is to remove xml file saving shortly after
landing this and only keep the xml file loading code for one release,
so we have a chance to convert any existing files we have.