memory-stack: fix sub-stack iteration
This fixes how we iterate the sub-stacks when we are looking for an
existing sub-stack that's large enough for the requested allocation
size. The code wasn't correctly adapted previously to consider that
RutLists are circular and so we need to make sure we stop iteration at
the true head of the list instead of looping back around to the current
sub-stack. This bug meant we would sometimes re-allocated sub-stacks
before the stack had been rewound.

Additionally this also changes the code to use for_each_safe when
freeing a stack since we will otherwise be accessing the list_node of
each freed sub-stack which may contain junk.