rply: Add ply_start for reading from memory
This adds a new ply_start function to rply.c to enable reading from a
memory buffer. Since rply was written to use buffered io this is
implemented as a thin shim that uses fmemopen so that we can get a
FILE *fp to access the memory and leave most of rply.c untouched.