Removes the CoglSwapChain api
The CoglSwapChain api made initializing a context more awkward than it
ought to be in the relatively common case where you want onscreen
framebuffers to have an alpha channel.

The api let you specify an explicit swap-chain length and also request
that the swap chain buffers should have an alpha component and then the
a CoglSwapChain could be used to construct a CoglOnscreenTemplate

This removes the CoglSwapChain object and apis and adds a
cogl_onscreen_template_set_has_alpha() api instead. Since we can't
normally control exactly how long the swap chain is even if we have api
to request a length this patch doesn't replace that capability.

One thing to note here is that this removes the ability to request a
single buffered CoglOnscreen framebuffer but since we have never had a
use case for that and have not generally considered the implications of
supporting single buffered rendering in Cogl's design it is probably for
the best not to imply that it's supported.

Reviewed-by: Neil Roberts <neil@linux.intel.com>