pb-c-rpc: check closures after client_failed callback
client_failed() in rig-protobuf-c-rpc.c would save off a pointer to the
closures associated with a client before calling an associated
error_handler callback and then dereference the saved pointer which
could have been invalidated in the error handler. This changes the code
to only check for closures after calling the error handler.