pb-c-rpc: Adds rig_pb_rpc_client_add_connect_handler
This adds rig_pb_rpc_client_add_connect_handler api to let us register a
callback with a PB_RPC_Client that gets called when the client connects
to a server. This is much more convenient to use than having to poll the
connection status with rig_pb_rpc_client_is_connected. This change also
fixes a problem with the way we handled polling before that resulted in
us starving the mainloop after connecting an editor to a slave.