engine: split more out or rig_engine_init
This splits out the code for creating the editor ui and some code for
creating a debug gradient texture from rig_engine_init. It also moves up
a few bits of initialization that don't depend on any ui having been
loaded.

The aim is to help clarify what parts of initialization relate to
loading a new document since we currently make some assumptions about
only loading a document at initialization time but there is code in
rig_engine_init that needs to be dealt with directly as part of document
loading instead.