box-layout: Fix the axis used for the ‘other’ preferred size
When the preferred size of the box layout is queried it was using the
same axis for the ‘other’ size as for the ‘main’ size. This was
causing the inspector to report the tallest property inspector as its
width which would make it very wide when an entity with a text
component was selected.