gaussian-blurrer: handle resizing correctly
The problem with resizing was that we didn't update the viewports and
orthographic coordinates of the RutCameras used by the
rut-gaussian-blurrer api.

This simplifies the RutGaussianBlurrer by removing the use or RutCamera
which was an unnecessary complexity when the framebuffers used are
entirely internal.

Since we use cogl_offscreen_new_to_texture to create our framebuffer we
can assume that the viewport is always correct, but we now also use
cogl_framebuffer_orthographic() to setup the coordinates.