rig-camera-view: add rig_camera_view_set_scene api
To enable reloading a new scene in Rig gracefully we need the ability to
destroy an entity scene, create a new one and point a RigCameraView at
the new scene. This adds a rig_camera_view_set_scene() function so that
the scene being viewed can be changed at runtime.