Add a seperate camera for play mode
Play mode now has a seperate entity to represent the camera. This is
automatically added to the scene if it doesn't already exist after
loading a file. The camera is not internal so it can be animated,
serialized and saved. The RigCameraView now uses this camera instead
of its view camera when in play mode. The entity has a handle entity
so that it can be selected and will be visible in the scene. However
it's currently quite difficult to find it because it is usually
positioned far away from the scene.