Reinstate the Ctrl+C key to select the camera
The Ctrl+C key used to select the view camera so that its properties
could be manipulated. This was removed since the switch to a separate
camera view in bf635357 made this difficult. This patch brings the
keybinding back but makes it select the global play camera instead of
the view camera. This is useful because it's not always easy to find
the handle for the play camera if it is far away from the scene.